move aws functionality to separate files
