+++ src/helpers.js
@@ -1,9 +1,18 @@
 #!/usr/bin/env node
 
 const _ = require("lodash")
+const fs = require("fs")
 const url = require("url")
 const path = require("path")
 const departmentsJson = require("./departments.json")
+
+const directoryExists = directory => {
+  return (
+    directory &&
+    fs.existsSync(directory) &&
+    fs.lstatSync(directory).isDirectory()
+  )
+}
 
 const findDepartmentByNumber = departmentNumber => {
   return departmentsJson.find(department => {
@@ -119,6 +128,7 @@ const pathToChildRecursive = (basePath, child, courseData) => {
 }
 
 module.exports = {
+  directoryExists,
   findDepartmentByNumber,
   getDepartments,
   getCourseNumbers,
